pyproject.toml
[project]
name = "pathwaykg"
version = "0.2.0"
description = "Build and plot RDF knowledge graphs of pathways from the KEGG database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "pyvis>=0.3.2",
    "rdflib>=7.6.0",
    "regex>=2026.2.28",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
]

[project.scripts]
kgbuild = "pathwaykg.build:main"
visualize = "pathwaykg.visualize:main"
